Fully implemented all new tools with models, recipes, etc.; added upgrade templates and removed legacy smithing recipes; added new 9-item amethyst block

+++ src/main/java/net/dollar/testmod/item/ModInfusedDiamondShovelItem.java
@@ -7,7 +7,7 @@ import net.minecraft.world.item.ShovelItem;
 import net.minecraft.world.item.Tier;
 
 public class ModInfusedDiamondShovelItem extends ShovelItem {
-    public ModInfusedDiamondShovelItem(Tier p_42961_, int p_42962_, float p_42963_, Properties p_42964_) {
+    public ModInfusedDiamondShovelItem(Tier p_42961_, float p_42962_, float p_42963_, Properties p_42964_) {
         super(p_42961_, p_42962_, p_42963_, p_42964_);
     }
